+++ examples/tests/dependency_injection/test_dependency_default_value_no_dependency_fn.py
@@ -1,6 +1,5 @@
-from starlette.status import HTTP_200_OK
-
 from examples.dependency_injection import dependency_default_value_no_dependency_fn
+from starlite.status_codes import HTTP_200_OK
 from starlite.testing import TestClient
