+++ examples/tests/dependency_injection/test_dependency_skip_validation.py
@@ -1,6 +1,5 @@
-from starlette.status import HTTP_200_OK
-
 from examples.dependency_injection import dependency_skip_validation
+from starlite.status_codes import HTTP_200_OK
 from starlite.testing import TestClient
