+++ examples/tests/dependency_injection/test_dependency_validation_error.py
@@ -1,6 +1,5 @@
-from starlette.status import HTTP_500_INTERNAL_SERVER_ERROR
-
 from examples.dependency_injection import dependency_validation_error
+from starlite.status_codes import HTTP_500_INTERNAL_SERVER_ERROR
 from starlite.testing import TestClient
