+++ tests/response/test_file_response.py
@@ -45,7 +45,7 @@ def test_file_response(tmpdir: Path) -> None:
         await response(scope, receive, send)
 
     assert filled_by_bg_task == ""
-    client = TestClient(app)  # type: ignore
+    client = TestClient(app)
     response = client.get("/")
     expected_disposition = 'attachment; filename="example.png"'
     assert response.status_code == HTTP_200_OK
@@ -75,7 +75,7 @@ def test_file_response_with_chinese_filename(tmpdir: Path) -> None:
     path = tmpdir / filename
     Path(path).write_bytes(content)
     app = FileResponse(path=path, filename=filename)
-    client = TestClient(app)  # type: ignore
+    client = TestClient(app)
     response = client.get("/")
     expected_disposition = "attachment; filename*=utf-8''%e4%bd%a0%e5%a5%bd.txt"
     assert response.status_code == HTTP_200_OK
@@ -89,7 +89,7 @@ def test_file_response_with_inline_disposition(tmpdir: Path) -> None:
     path = tmpdir / filename
     Path(path).write_bytes(content)
     app = FileResponse(path=path, filename=filename, content_disposition_type="inline")
-    client = TestClient(app)  # type: ignore
+    client = TestClient(app)
     response = client.get("/")
     expected_disposition = 'inline; filename="hello.txt"'
     assert response.status_code == HTTP_200_OK
@@ -102,6 +102,6 @@ def test_file_response_known_size(tmpdir: Path) -> None:
     content = b"<file content>" * 1000
     Path(path).write_bytes(content)
     app = FileResponse(path=path, filename="example.png")
-    client: TestClient = TestClient(app)  # type: ignore
+    client: TestClient = TestClient(app)
     response = client.get("/")
     assert response.headers["content-length"] == str(len(content))
